--- TeamCode/src/main/java/org/firstinspires/ftc/teamcode/SKIBIDI_TEST.java
@@ -1,17 +1,0 @@
-package org.firstinspires.ftc.teamcode;
-
-import com.qualcomm.robotcore.eventloop.opmode.OpMode;
-import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
-
-@TeleOp
-public class SKIBIDI_TEST extends OpMode {
-    @Override
-    public void init() {
-
-    }
-
-    @Override
-    public void loop() {
-        System.out.println("SKIBIDI 12");
-    }
-}
