+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/helpers/utils/opmodes/LimelightAimerTest.java
@@ -10,8 +10,8 @@ import com.pedropathing.follower.Follower;
 import com.pedropathing.localization.Pose;
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
 
-import org.firstinspires.ftc.teamcode.Pedro.constants.FConstants;
-import org.firstinspires.ftc.teamcode.Pedro.constants.LConstants;
+import org.firstinspires.ftc.teamcode.pedro.constants.FConstants;
+import org.firstinspires.ftc.teamcode.pedro.constants.LConstants;
 import org.firstinspires.ftc.teamcode.auto.sample.SubmersibleGrabV2;
 import org.firstinspires.ftc.teamcode.helpers.commands.RepeatUntilCommand;
 import org.firstinspires.ftc.teamcode.helpers.opmode.VLRLinearOpMode;
