+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/pedro/constants/LConstants.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.Pedro.constants;
+package org.firstinspires.ftc.teamcode.pedro.constants;
 
 import com.pedropathing.localization.GoBildaPinpointDriver;
 import com.pedropathing.localization.constants.PinpointConstants;
