+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/pedro/examples/ExampleBucketAuto.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.Pedro.examples;
+package org.firstinspires.ftc.teamcode.pedro.examples;
 
 import com.pedropathing.follower.Follower;
 import com.pedropathing.localization.Pose;
@@ -11,8 +11,8 @@ import com.pedropathing.util.Timer;
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
 import com.qualcomm.robotcore.eventloop.opmode.OpMode;
 
-import org.firstinspires.ftc.teamcode.Pedro.constants.FConstants;
-import org.firstinspires.ftc.teamcode.Pedro.constants.LConstants;
+import org.firstinspires.ftc.teamcode.pedro.constants.FConstants;
+import org.firstinspires.ftc.teamcode.pedro.constants.LConstants;
 
 /**
  * This is an example auto that showcases movement and control of two servos autonomously.
