+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/pedro/tuners_tests/automatic/ForwardVelocityTuner.java
@@ -1,13 +1,13 @@
-package org.firstinspires.ftc.teamcode.Pedro.tuners_tests.automatic;
+package org.firstinspires.ftc.teamcode.pedro.tuners_tests.automatic;
 
 import static com.pedropathing.follower.FollowerConstants.leftFrontMotorDirection;
-import static com.pedropathing.follower.FollowerConstants.leftFrontMotorName;
 import static com.pedropathing.follower.FollowerConstants.leftRearMotorDirection;
-import static com.pedropathing.follower.FollowerConstants.leftRearMotorName;
 import static com.pedropathing.follower.FollowerConstants.rightFrontMotorDirection;
-import static com.pedropathing.follower.FollowerConstants.rightFrontMotorName;
 import static com.pedropathing.follower.FollowerConstants.rightRearMotorDirection;
-import static com.pedropathing.follower.FollowerConstants.rightRearMotorName;
+import static org.firstinspires.ftc.teamcode.subsystems.chassis.ChassisConfiguration.MOTOR_LEFT_BACK;
+import static org.firstinspires.ftc.teamcode.subsystems.chassis.ChassisConfiguration.MOTOR_LEFT_FRONT;
+import static org.firstinspires.ftc.teamcode.subsystems.chassis.ChassisConfiguration.MOTOR_RIGHT_BACK;
+import static org.firstinspires.ftc.teamcode.subsystems.chassis.ChassisConfiguration.MOTOR_RIGHT_FRONT;
 
 import com.acmerobotics.dashboard.FtcDashboard;
 import com.acmerobotics.dashboard.config.Config;
@@ -22,19 +22,19 @@ import com.qualcomm.robotcore.hardware.DcMotorEx;
 import com.qualcomm.robotcore.hardware.configuration.typecontainers.MotorConfigurationType;
 
 import org.firstinspires.ftc.robotcore.external.Telemetry;
-import org.firstinspires.ftc.teamcode.Pedro.constants.FConstants;
-import org.firstinspires.ftc.teamcode.Pedro.constants.LConstants;
+import org.firstinspires.ftc.teamcode.pedro.constants.FConstants;
+import org.firstinspires.ftc.teamcode.pedro.constants.LConstants;
 
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 
 /**
- * This is the StrafeVelocityTuner autonomous follower OpMode. This runs the robot right at max
+ * This is the ForwardVelocityTuner autonomous follower OpMode. This runs the robot forwards at max
  * power until it reaches some specified distance. It records the most recent velocities, and on
  * reaching the end of the distance, it averages them and prints out the velocity obtained. It is
  * recommended to run this multiple times on a full battery to get the best results. What this does
- * is, when paired with ForwardVelocityTuner, allows FollowerConstants to create a Vector that
+ * is, when paired with StrafeVelocityTuner, allows FollowerConstants to create a Vector that
  * empirically represents the direction your mecanum wheels actually prefer to go in, allowing for
  * more accurate following.
  * You can adjust the distance the robot will travel on FTC Dashboard: 192/168/43/1:8080/dash
@@ -45,8 +45,8 @@ import java.util.List;
  * @version 1.0, 3/13/2024
  */
 @Config
-@Autonomous(name = "Strafe Velocity Tuner", group = "Automatic Tuners")
-public class StrafeVelocityTuner extends OpMode {
+@Autonomous(name = "Forward Velocity Tuner", group = "Automatic Tuners")
+public class ForwardVelocityTuner extends OpMode {
     private ArrayList<Double> velocities = new ArrayList<>();
 
     private DcMotorEx leftFront;
@@ -57,7 +57,7 @@ public class StrafeVelocityTuner extends OpMode {
 
     private PoseUpdater poseUpdater;
 
-    public static double DISTANCE = 48;
+    public static double DISTANCE = 36;
     public static double RECORD_NUMBER = 10;
 
     private Telemetry telemetryA;
@@ -71,13 +71,12 @@ public class StrafeVelocityTuner extends OpMode {
     @Override
     public void init() {
         FConstants.initialize();
-
         poseUpdater = new PoseUpdater(hardwareMap, FConstants.class, LConstants.class);
 
-        leftFront = hardwareMap.get(DcMotorEx.class, leftFrontMotorName);
-        leftRear = hardwareMap.get(DcMotorEx.class, leftRearMotorName);
-        rightRear = hardwareMap.get(DcMotorEx.class, rightRearMotorName);
-        rightFront = hardwareMap.get(DcMotorEx.class, rightFrontMotorName);
+        leftFront = hardwareMap.get(DcMotorEx.class, MOTOR_LEFT_FRONT);
+        leftRear = hardwareMap.get(DcMotorEx.class, MOTOR_LEFT_BACK);
+        rightRear = hardwareMap.get(DcMotorEx.class, MOTOR_RIGHT_BACK);
+        rightFront = hardwareMap.get(DcMotorEx.class, MOTOR_RIGHT_FRONT);
         leftFront.setDirection(leftFrontMotorDirection);
         leftRear.setDirection(leftRearMotorDirection);
         rightFront.setDirection(rightFrontMotorDirection);
@@ -100,28 +99,31 @@ public class StrafeVelocityTuner extends OpMode {
         }
 
         telemetryA = new MultipleTelemetry(this.telemetry, FtcDashboard.getInstance().getTelemetry());
-        telemetryA.addLine("The robot will run at 1 power until it reaches " + DISTANCE + " inches to the right.");
+        telemetryA.addLine("The robot will run at 1 power until it reaches " + DISTANCE + " inches forward.");
         telemetryA.addLine("Make sure you have enough room, since the robot has inertia after cutting power.");
-        telemetryA.addLine("After running the distance, the robot will cut power from the drivetrain and display the strafe velocity.");
+        telemetryA.addLine("After running the distance, the robot will cut power from the drivetrain and display the forward velocity.");
         telemetryA.addLine("Press CROSS or A on game pad 1 to stop.");
+        telemetryA.addData("pose", poseUpdater.getPose());
         telemetryA.update();
+
     }
 
     /**
-     * This starts the OpMode by setting the drive motors to run right at full power.
+     * This starts the OpMode by setting the drive motors to run forward at full power.
      */
     @Override
     public void start() {
         leftFront.setPower(1);
-        leftRear.setPower(-1);
-        rightFront.setPower(-1);
+        leftRear.setPower(1);
+        rightFront.setPower(1);
         rightRear.setPower(1);
+        end = false;
     }
 
     /**
      * This runs the OpMode. At any point during the running of the OpMode, pressing CROSS or A on
-     * game pad1 will stop the OpMode. This continuously records the RECORD_NUMBER most recent
-     * velocities, and when the robot has run sideways enough, these last velocities recorded are
+     * game pad 1 will stop the OpMode. This continuously records the RECORD_NUMBER most recent
+     * velocities, and when the robot has run forward enough, these last velocities recorded are
      * averaged and printed.
      */
     @Override
@@ -136,14 +138,14 @@ public class StrafeVelocityTuner extends OpMode {
 
         poseUpdater.update();
         if (!end) {
-            if (Math.abs(poseUpdater.getPose().getY()) > DISTANCE) {
+            if (Math.abs(poseUpdater.getPose().getX()) > DISTANCE) {
                 end = true;
                 for (DcMotorEx motor : motors) {
                     motor.setZeroPowerBehavior(DcMotor.ZeroPowerBehavior.BRAKE);
                     motor.setPower(0);
                 }
             } else {
-                double currentVelocity = Math.abs(MathFunctions.dotProduct(poseUpdater.getVelocity(), new Vector(1, Math.PI / 2)));
+                double currentVelocity = Math.abs(MathFunctions.dotProduct(poseUpdater.getVelocity(), new Vector(1, 0)));
                 velocities.add(currentVelocity);
                 velocities.remove(0);
             }
@@ -161,7 +163,7 @@ public class StrafeVelocityTuner extends OpMode {
             }
             average /= (double) velocities.size();
 
-            telemetryA.addData("strafe velocity:", average);
+            telemetryA.addData("forward velocity:", average);
             telemetryA.update();
         }
     }
