+++ TeamCode/src/main/res/xml/vlr_itd_v1.xml
@@ -1,0 +1,30 @@
+<?xml version='1.0' encoding='UTF-8' standalone='yes' ?>
+
+<Robot type="FirstInspires-FTC">
+    <LynxUsbDevice name="Control Hub Portal" serialNumber="(embedded)" parentModuleAddress="173">
+        <ServoHub name="Servo Hub 3" port="3">
+            <Servo name="angle" port="0" />
+            <Servo name="twist" port="4" />
+            <Servo name="claw" port="5" />
+            <ContinuousRotationServo name="leftHang" port="3" />
+            <ContinuousRotationServo name="rightHang" port="1" />
+        </ServoHub>
+        <LynxModule name="Control Hub" port="173">
+            <goBILDA5202SeriesMotor name="MotorRightBack" port="0" />
+            <goBILDA5202SeriesMotor name="MotorRightFront" port="1" />
+            <goBILDA5202SeriesMotor name="MotorLeftFront" port="2" />
+            <goBILDA5202SeriesMotor name="MotorLeftBack" port="3" />
+            <Servo name="ServoWiper" port="0" />
+            <RevBlinkinLedDriver name="blinkin" port="1" />
+            <RevSPARKMini name="MotorRotator" port="2" />
+            <RevSPARKMini name="MotorArm1" port="3" />
+            <RevSPARKMini name="MotorArm2" port="4" />
+            <RevSPARKMini name="MotorArm3" port="5" />
+            <RevTouchSensor name="SlideLimit" port="1" />
+            <AnalogInput name="leftAnalog" port="1" />
+            <AnalogInput name="rightAnalog" port="0" />
+            <goBILDAPinpoint name="pinpoint" port="0" bus="0" />
+        </LynxModule>
+    </LynxUsbDevice>
+    <EthernetDevice name="limelight" serialNumber="EthernetOverUsb:eth0:172.29.0.29" port="0" ipAddress="172.29.0.1" />
+</Robot>
