+++ TeamCode/src/main/res/xml/vlr_itd_v2.xml
@@ -1,0 +1,41 @@
+<?xml version='1.0' encoding='UTF-8' standalone='yes' ?>
+<Robot type="FirstInspires-FTC">
+    <LynxUsbDevice name="Expansion Hub Portal 1" serialNumber="DQ8OZVBG" parentModuleAddress="2">
+        <LynxModule name="Expansion Hub 2" port="2">
+            <goBILDA5202SeriesMotor name="MotorRotator" port="0" />
+            <goBILDA5202SeriesMotor name="MotorArm1" port="1" />
+            <goBILDA5202SeriesMotor name="MotorArm2" port="2" />
+            <goBILDA5202SeriesMotor name="MotorArm3" port="3" />
+            <RevTouchSensor name="SlideLimit" port="2" />
+            <AnalogInput name="clawAnalogProximity" port="2" />
+            <AnalogInput name="leftAnalog" port="0" />
+            <AnalogInput name="rightAnalog" port="1" />
+        </LynxModule>
+    </LynxUsbDevice>
+
+    <LynxUsbDevice name="Control Hub Portal" serialNumber="(embedded)" parentModuleAddress="173">
+        <ServoHub name="Servo Hub 3" port="3">
+            <Servo name="angle" port="1" />
+            <ContinuousRotationServo name="rightHang" port="4" />
+            <ContinuousRotationServo name="leftHang" port="5" />
+            <Servo name="twist" port="3" />
+            <Servo name="claw" port="2" />
+        </ServoHub>
+
+        <LynxModule name="Control Hub" port="173">
+            <goBILDA5202SeriesMotor name="MotorRightBack" port="3" />
+            <goBILDA5202SeriesMotor name="MotorRightFront" port="2" />
+            <goBILDA5202SeriesMotor name="MotorLeftFront" port="0" />
+            <goBILDA5202SeriesMotor name="MotorLeftBack" port="1" />
+            <Servo name="ServoWiper" port="1" />
+            <RevBlinkinLedDriver name="blinkin" port="5" />
+            <AnalogInput name="CHanalog0" port="0" />
+            <AnalogInput name="CHanalog1" port="1" />
+            <AnalogInput name="CHanalog2" port="2" />
+            <AnalogInput name="CHanalog3" port="3" />
+            <RevTouchSensor name="BeamBreak" port="1" />
+            <goBILDAPinpoint name="pinpoint" port="0" bus="1" />
+        </LynxModule>
+    </LynxUsbDevice>
+    <EthernetDevice name="limelight" serialNumber="EthernetOverUsb:eth0:172.29.0.29" port="0" ipAddress="172.29.0.1" />
+</Robot>
